# BookBackend/api/openAPI.yaml
openapi: "3.0.3"
info:
  title: "BookBackend API"
  description: "BookBackend API"
  version: "1.0.0"
servers:
  - url: "https://BookBackend"
paths:
  /media:
    get:
      tags:
        - "Media"
      summary: "GET media"
      operationId: "getAllMedia"
      responses:
        "200":
          $ref: "#/components/responses/MediaList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /media/findByType:
    get:
      tags:
        - "Media"
      summary: "GET media/findByType"
      operationId: "findByType"
      parameters:
        - name: "type"
          in: "query"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          $ref: "#/components/responses/MediaList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
...
  responses:
    MediaList:
      description: "OK"
      content:
        application/json:
          schema:
            anyOf:
              - type: "array"
                items:
                  $ref: "#/components/schemas/Book"
              - type: "array"
                items:
                  $ref: "#/components/schemas/Cd"
...
    BadRequest:
        description: "Bad Request"
        content:
            application/json:
              schema:
                  $ref: "#/components/schemas/Error"
    Unauthorized:
        description: "Unauthorized"
        content:
            application/json:
              schema:
                  $ref: "#/components/schemas/Error"
  schemas:
...
    Cd:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 1234
        title:
          type: "string"
          example: "The Lord of the Rings"
        discription:
          type: "string"
          example: "Der erste Band der Trilogie"
        artist:
          type: "string"
          example: "J.R.R. Tolkien"
        label:
          type: "string"
          example: "Hörbuch"
...
    Book:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 1234
        language:
          type: "string"
          example: "de"
        title:
          type: "string"
          example: "The Lord of the Rings"
        isbn:
          type: "string"
          example: "978-3-16-148410-0"
        releaseDate:
          type: "string"
          format: "date-time"
        discription:
          type: "string"
          example: "Der erste Band der Trilogie"
        publisher:
          type: "string"
          example: "xyz-publisher"
        author:
          type: array
          items:
            $ref: "#/components/schemas/Author"
        label:
          type: "string"
          example: "Hörbuch"
...
    Author:
        type: "object"
        properties:
            id:
              type: "integer"
              format: "int64"
            name:
              type: "string"
            surname:
              type: "string"
...
    Error:
        type: "object"
        properties:
            message:
              type: "string"
              example: "Bad Request"
  securitySchemes:
    prod:
      scheme: bearer
      type: http
      bearerFormat: JWT